edit Register::getData(int index)

+++ Task3/MLS.h
@@ -10,7 +10,7 @@ private:
 
 public:
     void setData(int index, int bitPattern); // setter
-    int8_t getData(int index);
+    int getData(int index);
     void reset();
     void show();
     int8_t add(int index, int index2);
